Modularized last commit
